Implement workaround required for Wayland on Linux
Similar issue which was met on endless-sky game. Fedora may prefer
wayland and emit glewInit() failure, but otherwise it is initialized
correctly. Detect that case and pass into the application. It should
work if SDL were correctly initialized.

Glut does not seem to be needed for anything in this case.

More on: https://github.com/endless-sky/endless-sky/pull/7027